Made a reusable component for [text + Pro badge]
